Create LICENSE
Adding detectable licence (see https://docs.github.com/en/github/building-a-strong-community/adding-a-license-to-a-repository). As far as I can see most other repos in Ensembl have implemented it. Further I also updated the Copyright section to be in sync with the rest of Ensembl.
 
I did not add these sentences any more:
"
The files in the /src and /parallel directories are under the following license.
Please look into the /samtools directory for its separate license.
"